Fix a breaking change in DuckTyping support
Pre-.NET 8, we use reflection to invoke the private member `MethodBuilder.GetToken()`.
However, in .NET 8 they moved a bunch of reflection emit APIs around:
https://github.com/dotnet/runtime/pull/78544
The net result is that the `GetToken()` method no longer exists. Luckily, the public
`MetadataToken` property directly calls `GetToken()`, so we can use that instead.

For the fix, we add a fallback for looking for that token. Theoretically, at some point in the future, we could switch the order (when/if .NET 8+ becomes the most common FX)