Build native parts using .NET 7 and Centos7
This is one of the trickiest parts of the upgrade. .NET 8 doesn't support CentOs but
we still do, so we _must_ compile the native parts in CentOS (where .NET 8 can't run).

As a workaround, we build all the native parts in Centos7 (using .NET 7), then do the
rest of the build (the managed parts) in the Debian image (where we can run .NET 8).

The final sequence is
- Build managed loader in Debian with .NET 8
- Build native loader + tracer in centos with .NET 7
- Build remaining managed code in Debian with .NET 8

For the profiler, we build everything in centos7