Update dd_dotnet to target .NET 8
This (somewhat surprisingly) seems to work ok, though it does have an element of risk.
As we're doing NativeAOT against .NET 8, inside Debian, theoretically we're linking
against a version of glibc that isn't supported in centos. However, smoke tests seem
to show that it works. So either are tests aren't correct, or we're in the clear.
Worst case, as we haven't officially announced this yet, hopefully we can address
the issue later if we need to